Migrate BookNewHearing
