+++ AdminWebsite/AdminWebsite.UnitTests/Controllers/HearingsController/BookNewHearingTests.cs
@@ -1,48 +1,23 @@
-using AdminWebsite.Services;
-using System.Linq;
 using System.Threading.Tasks;
 using AdminWebsite.Contracts.Requests;
-using AdminWebsite.Security;
 using AdminWebsite.UnitTests.Helper;
-using UserApi.Client;
-using UserApi.Contract.Requests;
-using UserApi.Contract.Responses;
-using Autofac.Extras.Moq;
-using BookingsApi.Client;
 using BookingsApi.Contract.V2.Requests;
 using Microsoft.AspNetCore.Mvc;
-using VideoApi.Contract.Responses;
 using JudiciaryParticipantRequest = AdminWebsite.Contracts.Requests.JudiciaryParticipantRequest;
 using LinkedParticipantType = AdminWebsite.Contracts.Enums.LinkedParticipantType;
 
 namespace AdminWebsite.UnitTests.Controllers.HearingsController
 {
-    public class BookNewHearingTests
+    public class BookNewHearingTests : HearingsControllerTests
     {
-        private AutoMock _mocker;
-        private AdminWebsite.Controllers.HearingsController _controller;
         private string _expectedUserIdentityName;
 
         [SetUp]
-        public void Setup()
+        protected override void Setup()
         {
+            base.Setup();
             _expectedUserIdentityName = "created by";
-
-            _mocker = AutoMock.GetLoose();
-            _mocker.Mock<IConferenceDetailsService>().Setup(cs => cs.GetConferenceDetailsByHearingId(It.IsAny<Guid>(), false))
-                .ReturnsAsync(new ConferenceDetailsResponse
-                {
-                    MeetingRoom = new MeetingRoomResponse
-                    {
-                        AdminUri = "AdminUri",
-                        JudgeUri = "JudgeUri",
-                        ParticipantUri = "ParticipantUri",
-                        PexipNode = "PexipNode",
-                        PexipSelfTestNode = "PexipSelfTestNode",
-                        TelephoneConferenceId = "expected_conference_phone_id"
-                    }
-                });
-            _controller = _mocker.Create<AdminWebsite.Controllers.HearingsController>();
+            UserIdentity.Setup(x => x.GetUserIdentityName()).Returns(_expectedUserIdentityName);
         }
         
         [Test]
@@ -64,14 +39,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 .WithParticipant("Judicial Office Holder", "fname4.lname4@hmcts.net")
                 .WithParticipant("Judge", "judge.fudge@hmcts.net");
             
-            _mocker.Mock<IBookingsApiClient>().Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
+            BookingsApiClient.Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
                 .ReturnsAsync(hearingDetailsResponse);
-
-            
-            _mocker.Mock<IUserIdentity>().Setup(x => x.GetUserIdentityName()).Returns(_expectedUserIdentityName);
             
             // Act
-            var result = await _controller.Post(bookingRequest);
+            var result = await Controller.Post(bookingRequest);
 
             // Assert
             result.Result.Should().BeOfType<CreatedResult>();
@@ -83,15 +55,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
             bookingDetails.Participants.Exists(x => string.IsNullOrWhiteSpace(x.Username)).Should().BeFalse();
 
             bookingDetails.CreatedBy.Should().Be(_expectedUserIdentityName);
-            
-            _mocker.Mock<IHearingsService>().Verify(x
-                => x.AssignEndpointDefenceAdvocates(It.IsAny<List<EndpointRequest>>(), 
-                    It.Is<IReadOnlyCollection<ParticipantRequest>>(x => x.SequenceEqual(bookingDetails.Participants.AsReadOnly()))), Times.Once);
 
-            _mocker.Mock<IBookingsApiClient>().Verify(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()), Times.Once);
+            BookingsApiClient.Verify(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()), Times.Once);
         }
 
-        private BookingDetailsRequest InitHearingForV2Test()
+        private static BookingDetailsRequest InitHearingForV2Test()
         {
             // request with existing person, new user, existing user in AD but not in persons table 
             var bookNewHearingRequest = new BookingDetailsRequest
@@ -148,9 +116,9 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 Endpoints = new List<EndpointRequest>
                 {
                     new()
-                        {DisplayName = "displayname1", DefenceAdvocateContactEmail = "username1@hmcts.net"},
+                        {DisplayName = "displayname1", DefenceAdvocateContactEmail = "contact2@hmcts.net"},
                     new()
-                        {DisplayName = "displayname2", DefenceAdvocateContactEmail = "fname2.lname2@hmcts.net"},
+                        {DisplayName = "displayname2", DefenceAdvocateContactEmail = "contact3@hmcts.net"},
                 },
                 LinkedParticipants = new List<LinkedParticipantRequest>
                 {
@@ -174,71 +142,6 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 }
             };
             
-            return SetUpRequest(bookNewHearingRequest);
-        }
-
-        private BookingDetailsRequest SetUpRequest(BookingDetailsRequest bookNewHearingRequest)
-        {
-            _mocker.Mock<IUserAccountService>().Setup(x => x.GetAdUserIdForUsername(It.IsAny<string>())).ReturnsAsync(Guid.NewGuid().ToString());
-
-            foreach (var participant in bookNewHearingRequest.Participants.Where(x =>
-                !string.IsNullOrWhiteSpace(x.Username)))
-            {
-                var profile = new UserProfile
-                {
-                    UserId = Guid.NewGuid().ToString(),
-                    UserName = participant.Username,
-                    FirstName = participant.FirstName,
-                    LastName = participant.LastName
-                };
-                _mocker.Mock<IUserApiClient>().Setup(x => x.GetUserByAdUserIdAsync(It.Is<string>(e => e == participant.Username)))
-                    .ReturnsAsync(profile);
-            }
-
-            foreach (var participant in bookNewHearingRequest.Participants.Where(x =>
-                string.IsNullOrWhiteSpace(x.Username)))
-            {
-                var newUser = new NewUserResponse()
-                {
-                    UserId = Guid.NewGuid().ToString(),
-                    Username = $"{participant.FirstName}.{participant.LastName}@hmcts.net",
-                    OneTimePassword = "randomTest123"
-                };
-                _mocker.Mock<IUserApiClient>()
-                    .Setup(x => x.CreateUserAsync(It.Is<CreateUserRequest>(userRequest =>
-                        userRequest.RecoveryEmail == participant.ContactEmail))).ReturnsAsync(newUser);
-            }
-
-            var existingPat3 = bookNewHearingRequest.Participants.Single(x => x.ContactEmail == "contact3@hmcts.net");
-
-            var existingUser3 = new UserProfile()
-            {
-                UserId = Guid.NewGuid().ToString(),
-                UserName = $"{existingPat3.FirstName}.{existingPat3.LastName}@hmcts.net",
-                Email = existingPat3.ContactEmail,
-                FirstName = existingPat3.FirstName,
-                LastName = existingPat3.LastName,
-                DisplayName = existingPat3.DisplayName,
-            };
-            
-            _mocker.Mock<IUserApiClient>()
-                .Setup(x => x.GetUserByEmailAsync(existingPat3.ContactEmail)).ReturnsAsync(existingUser3);
-
-            _mocker.Mock<IPollyRetryService>().Setup(x => x.WaitAndRetryAsync<Exception, Task>
-                (
-                    It.IsAny<int>(), It.IsAny<Func<int, TimeSpan>>(), It.IsAny<Action<int>>(),
-                    It.IsAny<Func<Task, bool>>(), It.IsAny<Func<Task<Task>>>()
-                ))
-                .Callback(async (int retries, Func<int, TimeSpan> sleepDuration, Action<int> retryAction,
-                    Func<Task, bool> handleResultCondition, Func<Task> executeFunction) =>
-                {
-                    sleepDuration(1);
-                    retryAction(1);
-                    handleResultCondition(Task.CompletedTask);
-                    await executeFunction();
-                })
-                .ReturnsAsync(Task.CompletedTask);
-
             return bookNewHearingRequest;
         }
     }
