--- AdminWebsite/AdminWebsite/Exceptions/HearingsServiceException.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace AdminWebsite.Exceptions;
-
-public class HearingsServiceException : Exception
-{
-    public HearingsServiceException(string message) : base(message)
-    {
-    }
-}
