+++ AdminWebsite/AdminWebsite/Exceptions/ServiceException.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace AdminWebsite.Exceptions;
+
+public class ServiceException(string message) : Exception(message);
